rl/requirements.txt
# RL Agent
torch==2.6.0
numpy==1.26.4
stable-baselines3[extra]==2.6.0
gymnasium[mujoco]==1.1.1
matplotlib==3.10.1
sb3-contrib==2.6.0
pyyaml==6.0.1
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
hydra-submitit-launcher==1.2.0
omegaconf==2.3.0
moviepy==2.1.2
mlflow==2.21.3
triton==3.2.0
